pyproject.toml
[tool.poetry]
name = "mulatu"
version = "0.1.0"
description = ""
authors = ["Hadrien David <hadrien@ectobal.com>"]

[tool.poetry.dependencies]
python = "^3.7"
structlog = "^20.1.0"

[tool.poetry.dev-dependencies]
pytest-watch = "^4.2.0"
pdbpp = "^0.10.0"
pylama = "^7.7.1"
mypy = "^0.720"
black = "^19.3b0"
python-semantic-release = "^4.4.0"
isort = "^4.3.21"
pytest-cov = "^2.8.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry.scripts]
mulatu = "mulatu:main"
